+++ src/gdmbayes/plotting/__init__.py
@@ -1,11 +1,16 @@
 """Plotting utilities for spGDMM models."""
 
+from .fit import plot_link_curve, plot_obs_vs_pred
+from .importance import plot_predictor_importance
 from .isplines import plot_isplines
 from .ppc import plot_ppc
 from .scoring import crps_boxplot
 
 __all__ = [
     "plot_isplines",
+    "plot_predictor_importance",
+    "plot_obs_vs_pred",
+    "plot_link_curve",
     "plot_ppc",
     "crps_boxplot",
 ]